refactor(app): Rework admin users component as functional components for testing
